requirements.txt
aiokafka==0.7.2
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
fastapi==0.70.1
findspark==1.4.2
h11==0.12.0
httptools==0.3.0
idna==3.3
kafka-python==2.0.2
oauthlib==3.1.1
pip==21.3.1
py4j==0.10.9.2
pydantic==1.9.0
pyspark==3.2.0
python-dotenv==0.19.2
PyYAML==6.0
requests-oauthlib==1.3.0
requests==2.27.1
setuptools==57.0.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
watchgod==0.7
websockets==10.1
wheel==0.36.2
